+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.5] - 2026-07-24
+
 ### Added
 
 - **Search tables.** Schemas (PostgreSQL) and databases (MySQL) now carry a